Documentation and formatting

+++ initialsetup.sh
@@ -30,13 +30,16 @@ fi
       echo "We can restrict users to their own directory, or give them access to the whole system"
       read -p "Would you like to chroot/jail your users? [y/n] " jail
       case $jail in 
-            y)
-            echo ""
-            echo ""
-		read -p "What is the group name that you want for ftp users? " groupname
+      y)
+          echo ""
+          echo ""
+		      read -p "What is the group name that you want for ftp users? " groupname
 
-			addgroup --system $groupname
-			number=`grep -n "Subsystem" /etc/ssh/sshd_config | cut -d ":" -f1`
+			    addgroup --system $groupname
+
+          #Any other subsystem ftp stuff can conflict with what we want
+          #This will delete anything with Subsystem in it
+			    number=`grep -n "Subsystem" /etc/ssh/sshd_config | cut -d ":" -f1`
       		sed -i "${number}d" /etc/ssh/sshd_config
       		echo "Subsystem sftp internal-sftp" >> /etc/ssh/sshd_config
       		echo "Match Group $groupname" >> /etc/ssh/sshd_config
@@ -45,31 +48,36 @@ fi
       		echo "AllowTcpForwarding no" >> /etc/ssh/sshd_config
       		echo "ForceCommand internal-sftp" >> /etc/ssh/sshd_config
       		
-      		service ssh restart
-                  ;;
-            n)
-            echo ""
-            echo ""
-            read -p "What is the group name that you want for ftp users? " groupname
-                  addgroup --system $groupname
-                  number=`grep -n "Subsystem" /etc/ssh/sshd_config | cut -d ":" -f1`
-                  sed -i "${number}d" /etc/ssh/sshd_config
-                  echo "Subsystem sftp internal-sftp" >> /etc/ssh/sshd_config
-                  echo "Match Group $groupname" >> /etc/ssh/sshd_config
-                  echo "X11Forwarding no" >> /etc/ssh/sshd_config
-                  echo "AllowTcpForwarding no" >> /etc/ssh/sshd_config
-                  echo "ForceCommand internal-sftp" >> /etc/ssh/sshd_config
+      		  service ssh restart
+            ;;
+      n)
+          echo ""
+          echo ""
+          read -p "What is the group name that you want for ftp users? " groupname
+          addgroup --system $groupname
+          #Any other subsystem ftp stuff can conflict with what we want
+          #This will delete anything with Subsystem in it
+          number=`grep -n "Subsystem" /etc/ssh/sshd_config | cut -d ":" -f1`
+          sed -i "${number}d" /etc/ssh/sshd_config
+          echo "Subsystem sftp internal-sftp" >> /etc/ssh/sshd_config
+          echo "Match Group $groupname" >> /etc/ssh/sshd_config
+          echo "X11Forwarding no" >> /etc/ssh/sshd_config
+          echo "AllowTcpForwarding no" >> /etc/ssh/sshd_config
+          echo "ForceCommand internal-sftp" >> /etc/ssh/sshd_config
                   
-                  service ssh restart
+            service ssh restart
             ;;
       esac
+
       read -p "Do you wish to add a user? [y/n] " user
+      
       case $user in 
-            y)
- apt-get install wget;
-    
-    apt-get update
-    apt-get install php5-gd libssh2-php
+      y)
+ 
+          apt-get install wget;
+          apt-get update
+          apt-get install php5-gd libssh2-php
+          
 echo "First we need to figure out how we're adding an user(s)"
 echo ""
 echo ""
@@ -112,37 +120,47 @@ case $choice in
       echo "The file name and location you gave me was $dir/$file"
       echo "The password you gave me was $passwd"
       read -p "Is this correct? [y/n] " loop
-      if [ "$loop" = 'y' ];
+     if [ "$loop" = 'y' ];
         then
             for NAME in $NAMES; do
               
+              #Make the directories for the users
               mkdir /home/$NAME
               mkdir -p /var/www/$NAME
               mkdir /home/$NAME/public_html 
               mkdir /var/www/$NAME/public_html  
 
+              #Add the users and secure the crap out of them
               useradd -d /home/$NAME $NAME
               usermod -G $groupname $NAME
               usermod -s /bin/false $NAME
 
+              #Grab the latest version of wordpress and extract
               wget http://wordpress.org/latest.tar.gz -P /var/www/$NAME/
               cd /var/www/$NAME/
               tar xzvf latest.tar.gz
-              echo "$NAME:$passwd" | chpasswd         
+
+              #Change the password of the user
+              echo "$NAME:$passwd" | chpasswd  
+
           if [ "$jail" = 'y' ];
             then
+              #If the users are jailed root needs to own their directory
               chown root:root /home/$NAME
           fi
               chmod 0755 /home/$NAME
               cd /var/www/$NAME/
               cp -avr wordpress/ /var/www/$NAME/public_html/
              
-              #rm latest.tar.gz
-              #rm -rf wordpress/
+              #Give ownership
               chmod -R 755 * 
               chown $NAME:$groupname *
+
+              #To CHROOT the users we need to put in a folder that their account owns
+              #We partially escape the CHROOT by binding to a folder in their home directory
               echo "/var/www/$NAME/public_html /home/$NAME/public_html none bind 0 0" >> /etc/fstab
 
+              #Some wordpress php database config info
               echo "<?php" >> /var/www/$NAME/public_html/wordpress/wp-config.php
               echo "define('DB_NAME', '$NAME');" >> /var/www/$NAME/public_html/wordpress/wp-config.php
               echo "define('DB_USER', '$NAME');" >> /var/www/$NAME/public_html/wordpress/wp-config.php
@@ -156,6 +174,7 @@ case $choice in
               echo "define('ABSPATH', dirname(__FILE__) . '/');" >> /var/www/$NAME/public_html/wordpress/wp-config.php
               echo "require_once(ABSPATH . 'wp-settings.php');" >> /var/www/$NAME/public_html/wordpress/wp-config.php
 
+              #A .sql for granting permissions to a wordpress database
               echo "CREATE DATABASE $NAME;" >> name.sql
               echo "CREATE USER $NAME@localhost IDENTIFIED BY '$passwd';" >> name.sql
               echo "GRANT ALL PRIVILEGES ON $NAME.* TO $NAME@localhost;" >> name.sql
@@ -163,54 +182,63 @@ case $choice in
               echo "exit" >> name.sql
               mysql -u "root" -p$rootpasswd < name.sql
               
+              #Apply permissions to the directories
               chown -R $NAME:www-data *
               mkdir /var/www/$NAME/public_html/wordpress/wp-content/uploads
               chown -R :www-data /var/www/$NAME/public_html/wordpress/wp-content/uploads
+
+              #Cleaning up
               rm name.sql
               rm /var/www/$NAME/latest.tar.gz  
               rm -rf /var/www/$NAME/wordpress
     done
-      fi
-      ;;
-      2)
-            clear
-            read -p "What is the name of the user you wish to add? " NAME
-            echo "Now we need to set the users password."
-            echo ""
-            echo ""
-            read -p "What would you like the user(s) passwords to be? " passwd
-              read -p "What is the root MySQL password? " rootpasswd
-            echo ""
-            read -p "The password you entered was $passwd. Is this correct? [y/n] " loop
-      if [ "$loop" = 'y' ]
+  fi
+  ;;
+  2)
+    clear
+    read -p "What is the name of the user you wish to add? " NAME
+    echo ""
+    echo ""
+    read -p "What would you like the user(s) passwords to be? " passwd
+    echo "The password you gave me was $passwd"
+    echo ""
+       if [ "$loop" = 'y' ]
         then
-              
+              #Make the directories for the users
               mkdir /home/$NAME
               mkdir -p /var/www/$NAME
               mkdir /home/$NAME/public_html 
               mkdir /var/www/$NAME/public_html  
 
+              #Add the users and secure the crap out of them
               useradd -d /home/$NAME $NAME
               usermod -G $groupname $NAME
               usermod -s /bin/false $NAME
 
+              #Grab the latest version of wordpress and extract
               wget http://wordpress.org/latest.tar.gz -P /var/www/$NAME/
               cd /var/www/$NAME/
               tar xzvf latest.tar.gz
+              
+              #Change the password of the user
               echo "$NAME:$passwd" | chpasswd         
+          
           if [ "$jail" = 'y' ];
             then
+              #If the users are jailed root needs to own their directory
               chown root:root /home/$NAME
           fi
               chmod 0755 /home/$NAME
               cd /var/www/$NAME/
               cp -avr wordpress/ /var/www/$NAME/public_html/
-              #rm latest.tar.gz
-              #rm -rf wordpress/
+              #Give ownership
               chmod -R 755 * 
               chown $NAME:$groupname *
+              #To CHROOT the users we need to put in a folder that their account owns
+              #We partially escape the CHROOT by binding to a folder in their home directory
               echo "/var/www/$NAME/public_html /home/$NAME/public_html none bind 0 0" >> /etc/fstab
 
+              #Some wordpress php database config info
               echo "<?php" >> /var/www/$NAME/public_html/wordpress/wp-config.php
               echo "define('DB_NAME', '$NAME');" >> /var/www/$NAME/public_html/wordpress/wp-config.php
               echo "define('DB_USER', '$NAME');" >> /var/www/$NAME/public_html/wordpress/wp-config.php
@@ -224,6 +252,7 @@ case $choice in
               echo "define('ABSPATH', dirname(__FILE__) . '/');" >> /var/www/$NAME/public_html/wordpress/wp-config.php
               echo "require_once(ABSPATH . 'wp-settings.php');" >> /var/www/$NAME/public_html/wordpress/wp-config.php
 
+              #A .sql for granting permissions to a wordpress database
               echo "CREATE DATABASE $NAME;" >> name.sql
               echo "CREATE USER $NAME@localhost IDENTIFIED BY '$passwd';" >> name.sql
               echo "GRANT ALL PRIVILEGES ON $NAME.* TO $NAME@localhost;" >> name.sql
@@ -231,12 +260,17 @@ case $choice in
               echo "exit" >> name.sql
               mysql -u "root" -p$rootpasswd < name.sql
               
+              #Apply permissions to the directories
               chown -R $NAME:www-data *
               mkdir /var/www/$NAME/public_html/wordpress/wp-content/uploads
               chown -R :www-data /var/www/$NAME/public_html/wordpress/wp-content/uploads
+
+              #Cleaning up
               rm name.sql
+              rm /var/www/$NAME/latest.tar.gz  
+              rm -rf /var/www/$NAME/wordpress
       fi
-      ;;
+  ;;
 esac
             ;;
             n)
